requirements.txt
playwright
pyyaml
cf_clearance
